backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Data validation
pydantic==2.11.3
pydantic-settings==2.8.1

# AWS / Bedrock
boto3==1.37.36
botocore==1.37.36

# Environment variables
python-dotenv==1.1.0

# HTTP client (for any outbound requests)
httpx==0.28.1

# JSON handling (stdlib is fine, but this adds speed)
orjson==3.10.16
